Handle no infra/account/*.tfbackend files better for initial setup
Rather than erroring on a project with no `infra/account/*.tfbackend`
files, have `bin/accounts-ids-by-name` return a valid but empty JSON
object. This is more predictable behavior and allows Terraform to error
on the specific lookup locations rather than when the script is
triggered (which also avoids having to conditionally execute the script,
etc.).