docs: Add `Storage Blob Data Contributor` to admin permission list
At minimum only need this for the Terraform state storage account, but
for simplicity and generic usefulness of being able to access things in
storage during development, add it here.

The set-up-account script could assign the role scoped to that storage
account (after it creates it) to the person running things, or pre-check
access and provide nicer error message(s). But again, keeping things
simple right now.